Deprecated/requirements.txt
langchain>=0.1.0
langchain-community>=0.0.10
langchain_huggingface>=0.1.2
langchain-neo4j>=0.0.5
langchain_core==0.3.44
langchain_groq==0.2.5
python-dotenv==1.0.1
openai>=1.12.0
python-dotenv>=1.0.0
lightrag-hku==1.2.5